feat: add generate_dashboard service
Registers a givenergy_local.generate_dashboard service that auto-detects
all connected inverters and batteries, generates a Lovelace dashboard YAML
using the actual serial numbers, and writes it to the HA config directory.
A persistent notification shows the file path on completion.

Extracts the generation logic into custom_components/givenergy_local/lovelace.py
so it is shared by both the service and the standalone lovelace/generate.py script.
