+++ custom_components/givenergy_local/const.py
@@ -14,3 +14,4 @@ PLATFORMS = ["sensor", "switch", "number", "select", "time"]
 
 SERVICE_REBOOT_INVERTER = "reboot_inverter"
 SERVICE_CALIBRATE_BATTERY_SOC = "calibrate_battery_soc"
+SERVICE_GENERATE_DASHBOARD = "generate_dashboard"
