+++ lovelace/generate.py
@@ -20,11 +20,19 @@ Usage:
 from __future__ import annotations
 
 import argparse
+import importlib.util
 import sys
 from pathlib import Path
 
+# Load lovelace.py directly to avoid triggering the HA component's __init__.py.
+_spec = importlib.util.spec_from_file_location(
+    "givenergy_local_lovelace",
+    Path(__file__).parent.parent / "custom_components" / "givenergy_local" / "lovelace.py",
+)
+_mod = importlib.util.module_from_spec(_spec)  # type: ignore[arg-type]
+_spec.loader.exec_module(_mod)  # type: ignore[union-attr]
+generate_dashboard = _mod.generate_dashboard
 
-# ── entry point ───────────────────────────────────────────────────────────────
 
 def main() -> None:
     p = argparse.ArgumentParser(
@@ -39,448 +47,13 @@ def main() -> None:
                    help="Write to FILE instead of stdout")
     args = p.parse_args()
 
-    dashboard = _build(args.inverter, args.battery)
+    dashboard = generate_dashboard(args.inverter, args.battery)
 
     if args.output:
         Path(args.output).write_text(dashboard)
         print(f"Written to {args.output}", file=sys.stderr)
     else:
         sys.stdout.write(dashboard)
-
-
-# ── dashboard assembly ────────────────────────────────────────────────────────
-
-def _build(inv: str, bats: list[str]) -> str:
-    views = "\n".join([
-        _overview_view(inv),
-        _energy_view(inv),
-        _battery_view(bats),
-        _controls_view(inv),
-        _diagnostics_view(inv),
-    ])
-    return (
-        f"# Generated by lovelace/generate.py\n"
-        f"# Inverter : {inv}\n"
-        f"# Batteries: {', '.join(bats)}\n"
-        f"#\n"
-        f"# Required custom cards (install via HACS → Frontend):\n"
-        f"#   - power-flow-card-plus  (flixlix/power-flow-card-plus)\n"
-        f"#   - mini-graph-card  (kalkih/mini-graph-card)\n"
-        f"#   - apexcharts-card  (RomRider/apexcharts-card)\n"
-        f"\n"
-        f"title: GivEnergy\n"
-        f"views:\n"
-        f"\n"
-        f"{views}\n"
-    )
-
-
-# ── view builders ─────────────────────────────────────────────────────────────
-
-def _overview_view(inv: str) -> str:
-    return f"""\
-  # ── Overview ──────────────────────────────────────────────────────────────
-  - title: Overview
-    path: overview
-    icon: mdi:solar-power-variant
-
-    cards:
-      - type: custom:power-flow-card-plus
-        entities:
-          solar:
-            entity: {_i(inv, "pv_power")}
-            display_zero_state: true
-          battery:
-            entity: {_i(inv, "battery_power")}
-            state_of_charge: {_i(inv, "battery_soc")}
-          grid:
-            entity: {_i(inv, "grid_export_power")}
-          home:
-            entity: {_i(inv, "load_power")}
-
-      - type: glance
-        title: Status
-        columns: 4
-        entities:
-          - entity: {_i(inv, "status")}
-            name: Inverter
-          - entity: {_i(inv, "battery_soc")}
-            name: Battery SOC
-          - entity: {_i(inv, "battery_pause_mode")}
-            name: Pause Mode
-          - entity: {_i(inv, "battery_temperature")}
-            name: Battery Temp
-
-      - type: glance
-        title: Today
-        columns: 6
-        entities:
-          - entity: {_i(inv, "pv_energy_today")}
-            name: PV
-          - entity: {_i(inv, "battery_charge_today")}
-            name: Charged
-          - entity: {_i(inv, "battery_discharge_today")}
-            name: Discharged
-          - entity: {_i(inv, "grid_import_today")}
-            name: Imported
-          - entity: {_i(inv, "grid_export_today")}
-            name: Exported
-          - entity: {_i(inv, "load_energy_today")}
-            name: Consumed
-
-      - type: custom:apexcharts-card
-        header:
-          show: true
-          title: Power — Last 24 Hours
-        graph_span: 24h
-        series:
-          - entity: {_i(inv, "pv_power")}
-            name: PV
-            color: "#FFB300"
-          - entity: {_i(inv, "battery_power")}
-            name: Battery
-            color: "#42A5F5"
-          - entity: {_i(inv, "grid_export_power")}
-            name: Grid
-            color: "#66BB6A"
-          - entity: {_i(inv, "load_power")}
-            name: Load
-            color: "#EF5350"
-"""
-
-
-def _energy_view(inv: str) -> str:
-    def col_series(entity: str, name: str, color: str) -> str:
-        return (
-            f"          - entity: {entity}\n"
-            f"            name: {name}\n"
-            f"            color: \"{color}\"\n"
-            f"            type: column\n"
-            f"            statistics:\n"
-            f"              type: max\n"
-            f"              period: day"
-        )
-
-    return f"""\
-  # ── Energy ────────────────────────────────────────────────────────────────
-  - title: Energy
-    path: energy
-    icon: mdi:lightning-bolt
-
-    cards:
-      - type: custom:apexcharts-card
-        header:
-          show: true
-          title: Daily Generation vs Consumption — Last 30 Days
-        graph_span: 30d
-        series:
-{col_series(_i(inv, "pv_energy_today"), "PV Generated", "#FFB300")}
-{col_series(_i(inv, "load_energy_today"), "Consumed", "#EF5350")}
-
-      - type: custom:apexcharts-card
-        header:
-          show: true
-          title: Grid Import vs Export — Last 30 Days
-        graph_span: 30d
-        series:
-{col_series(_i(inv, "grid_export_today"), "Exported", "#66BB6A")}
-{col_series(_i(inv, "grid_import_today"), "Imported", "#EF5350")}
-
-      - type: custom:apexcharts-card
-        header:
-          show: true
-          title: Battery Charge vs Discharge — Last 30 Days
-        graph_span: 30d
-        series:
-{col_series(_i(inv, "battery_charge_today"), "Charged", "#42A5F5")}
-{col_series(_i(inv, "battery_discharge_today"), "Discharged", "#7E57C2")}
-
-      - type: entities
-        title: All-Time Totals
-        entities:
-          - entity: {_i(inv, "pv_energy_total")}
-            name: PV Generated
-          - entity: {_i(inv, "battery_throughput_total")}
-            name: Battery Throughput
-          - entity: {_i(inv, "grid_export_total")}
-            name: Grid Exported
-          - entity: {_i(inv, "grid_import_total")}
-            name: Grid Imported
-          - entity: {_i(inv, "inverter_output_total")}
-            name: Inverter Output
-          - entity: {_i(inv, "charge_from_grid_total")}
-            name: Charged from Grid
-          - entity: {_i(inv, "battery_discharge_this_year")}
-            name: Discharged This Year
-"""
-
-
-def _battery_view(bats: list[str]) -> str:
-    sections = "\n".join(_battery_section(s) for s in bats)
-    return f"""\
-  # ── Battery ───────────────────────────────────────────────────────────────
-  - title: Battery
-    path: battery
-    type: sections
-    icon: mdi:battery-high
-
-    sections:
-{sections}"""
-
-
-def _battery_section(serial: str) -> str:
-    pack = _bat_entity_rows(serial, [
-        ("soc",                  "SOC"),
-        ("voltage",              "Voltage"),
-        ("temperature_max",      "Temp Max"),
-        ("temperature_min",      "Temp Min"),
-        ("bms_mosfet_temperature", "BMS MOSFET Temp"),
-        None,
-        ("charge_cycles",        "Charge Cycles"),
-        ("remaining_capacity",   "Remaining Capacity"),
-        ("calibrated_capacity",  "Calibrated Capacity"),
-        ("design_capacity",      "Design Capacity"),
-        ("cell_voltages_sum",    "Cell Voltages Sum"),
-        ("cell_count",           "Cell Count"),
-    ])
-    cells = _bat_entity_rows(serial, [
-        (f"cell_{i}_voltage", f"Cell {i}") for i in range(1, 17)
-    ])
-    temps = _bat_entity_rows(serial, [
-        ("cells_1_4_temperature",   "Cells 1–4"),
-        ("cells_5_8_temperature",   "Cells 5–8"),
-        ("cells_9_12_temperature",  "Cells 9–12"),
-        ("cells_13_16_temperature", "Cells 13–16"),
-    ])
-
-    return f"""\
-      # ── {serial} ──
-      - cards:
-          - type: gauge
-            entity: {_b(serial, "soc")}
-            name: {serial}
-            min: 0
-            max: 100
-            needle: true
-            severity:
-              red: 0
-              yellow: 20
-              green: 40
-
-          - type: entities
-            title: Pack Details
-            entities:
-{pack}
-
-          - type: entities
-            title: Cell Voltages
-            entities:
-{cells}
-
-          - type: entities
-            title: Cell Temperatures
-            entities:
-{temps}
-"""
-
-
-def _controls_view(inv: str) -> str:
-    return f"""\
-  # ── Controls ──────────────────────────────────────────────────────────────
-  - title: Controls
-    path: controls
-    icon: mdi:tune
-
-    cards:
-      - type: entities
-        title: Mode
-        entities:
-          - entity: select.givenergy_inverter_{inv}_battery_power_mode
-            name: Battery Power Mode
-          - entity: {_i(inv, "battery_pause_mode")}
-            name: Pause Mode
-          - entity: {_i(inv, "battery_calibration_stage")}
-            name: Calibration Stage
-
-      - type: entities
-        title: Charging
-        entities:
-          - entity: switch.givenergy_inverter_{inv}_enable_charge
-            name: Enable Charge
-          - entity: number.givenergy_inverter_{inv}_charge_target_soc
-            name: Charge Target SOC
-          - entity: number.givenergy_inverter_{inv}_battery_soc_reserve
-            name: SOC Reserve
-          - entity: number.givenergy_inverter_{inv}_battery_charge_limit
-            name: Charge Power Limit
-          - type: divider
-          - entity: time.givenergy_inverter_{inv}_charge_slot_1_start
-            name: Slot 1 Start
-          - entity: time.givenergy_inverter_{inv}_charge_slot_1_end
-            name: Slot 1 End
-          - entity: time.givenergy_inverter_{inv}_charge_slot_2_start
-            name: Slot 2 Start
-          - entity: time.givenergy_inverter_{inv}_charge_slot_2_end
-            name: Slot 2 End
-
-      - type: entities
-        title: Discharging
-        entities:
-          - entity: switch.givenergy_inverter_{inv}_enable_discharge
-            name: Enable Discharge
-          - entity: number.givenergy_inverter_{inv}_battery_discharge_limit
-            name: Discharge Power Limit
-          - entity: number.givenergy_inverter_{inv}_battery_discharge_min_power_reserve
-            name: Min Power Reserve
-          - type: divider
-          - entity: time.givenergy_inverter_{inv}_discharge_slot_1_start
-            name: Slot 1 Start
-          - entity: time.givenergy_inverter_{inv}_discharge_slot_1_end
-            name: Slot 1 End
-          - entity: time.givenergy_inverter_{inv}_discharge_slot_2_start
-            name: Slot 2 Start
-          - entity: time.givenergy_inverter_{inv}_discharge_slot_2_end
-            name: Slot 2 End
-"""
-
-
-def _diagnostics_view(inv: str) -> str:
-    return f"""\
-  # ── Diagnostics ───────────────────────────────────────────────────────────
-  - title: Diagnostics
-    path: diagnostics
-    icon: mdi:wrench
-
-    cards:
-      - type: entities
-        title: Integration Health
-        entities:
-          - entity: {_i(inv, "last_successful_refresh")}
-            name: Last Successful Refresh
-          - entity: {_i(inv, "consecutive_refresh_failures")}
-            name: Consecutive Failures
-          - entity: {_i(inv, "total_refresh_failures")}
-            name: Total Failures
-
-      - type: entities
-        title: Faults & Warnings
-        entities:
-          - entity: {_i(inv, "status")}
-            name: Inverter Status
-          - entity: {_i(inv, "fault_code")}
-            name: Fault Code
-          - entity: {_i(inv, "fault_messages")}
-            name: Fault Messages
-          - entity: {_i(inv, "inverter_errors")}
-            name: Inverter Errors
-          - entity: {_i(inv, "charger_warning_code")}
-            name: Charger Warning Code
-          - entity: {_i(inv, "charge_status")}
-            name: Charge Status (raw)
-          - entity: {_i(inv, "system_mode")}
-            name: System Mode (raw)
-
-      - type: entities
-        title: Temperatures
-        entities:
-          - entity: {_i(inv, "battery_temperature")}
-            name: Battery
-          - entity: {_i(inv, "inverter_heatsink_temperature")}
-            name: Inverter Heatsink
-          - entity: {_i(inv, "charger_temperature")}
-            name: Charger
-
-      - type: entities
-        title: Electrical
-        entities:
-          - entity: {_i(inv, "ac_voltage")}
-            name: AC Voltage
-          - entity: {_i(inv, "ac_frequency")}
-            name: AC Frequency
-          - entity: {_i(inv, "battery_voltage")}
-            name: Battery Voltage
-          - entity: {_i(inv, "battery_current")}
-            name: Battery Current
-          - entity: {_i(inv, "grid_port_current")}
-            name: Grid Port Current
-          - entity: {_i(inv, "positive_dc_bus_voltage")}
-            name: Positive DC Bus
-          - entity: {_i(inv, "negative_dc_bus_voltage")}
-            name: Negative DC Bus
-          - entity: {_i(inv, "grid_apparent_power")}
-            name: Grid Apparent Power
-          - entity: {_i(inv, "inverter_power_factor")}
-            name: Inverter Power Factor
-
-      - type: entities
-        title: PV Strings
-        entities:
-          - entity: {_i(inv, "pv_string_1_voltage")}
-            name: String 1 Voltage
-          - entity: {_i(inv, "pv_string_1_current")}
-            name: String 1 Current
-          - entity: {_i(inv, "pv_string_1_power")}
-            name: String 1 Power
-          - entity: {_i(inv, "pv_string_2_voltage")}
-            name: String 2 Voltage
-          - entity: {_i(inv, "pv_string_2_current")}
-            name: String 2 Current
-          - entity: {_i(inv, "pv_string_2_power")}
-            name: String 2 Power
-
-      - type: entities
-        title: Hardware & Firmware
-        entities:
-          - entity: {_i(inv, "arm_firmware_version")}
-            name: ARM Firmware
-          - entity: {_i(inv, "dsp_firmware_version")}
-            name: DSP Firmware
-          - entity: {_i(inv, "modbus_version")}
-            name: Modbus Version
-          - entity: {_i(inv, "work_time_total")}
-            name: Work Time
-          - entity: {_i(inv, "device_type_code")}
-            name: Device Type Code
-          - entity: {_i(inv, "mppt_count")}
-            name: MPPT Count
-          - entity: {_i(inv, "phase_count")}
-            name: Phase Count
-          - entity: {_i(inv, "battery_type")}
-            name: Battery Type
-          - entity: {_i(inv, "meter_type")}
-            name: Meter Type
-          - entity: {_i(inv, "usb_device")}
-            name: USB Device
-          - entity: {_i(inv, "battery_nominal_capacity")}
-            name: Nominal Capacity (kWh)
-          - entity: {_i(inv, "battery_capacity")}
-            name: Capacity (Ah)
-"""
-
-
-# ── helpers ───────────────────────────────────────────────────────────────────
-
-def _i(serial: str, key: str) -> str:
-    """Inverter sensor entity ID."""
-    return f"sensor.givenergy_inverter_{serial}_{key}"
-
-
-def _b(serial: str, key: str) -> str:
-    """Battery sensor entity ID."""
-    return f"sensor.givenergy_battery_{serial}_{key}"
-
-
-def _bat_entity_rows(serial: str, items: list[tuple[str, str] | None]) -> str:
-    """Return indented YAML rows for a battery entities card."""
-    lines: list[str] = []
-    for item in items:
-        if item is None:
-            lines.append("              - type: divider")
-        else:
-            key, name = item
-            lines.append(f"              - entity: {_b(serial, key)}")
-            lines.append(f"                name: {name}")
-    return "\n".join(lines)
 
 
 if __name__ == "__main__":
